separate out the consts needed in airgap into its own file
